Final changes to functional code with cleanup
