+++ lib/shapes.js
@@ -7,7 +7,7 @@ class Circle {
     }
 
     render() {
-        return `<circle cx="${this.cx}" cy="${this.cy}" r="${this.r}"/><text font-family="Calibri" x="50" y="65" font-size="48" font-weight="bold" fill="#FFF" stroke="black" text-anchor="middle">`;
+        return ` width="300" height="200" viewBox="0 0 100 100" xmlns="http://www.w3.org/2000/svg"><circle cx="${this.cx}" cy="${this.cy}" r="${this.r}"/><text font-family="Calibri" x="50" y="65" font-size="48" font-weight="bold" fill="#FFF" stroke="black" text-anchor="middle">`;
     }
 }
 
@@ -21,7 +21,7 @@ class Square {
 
     }
     render(){
-        return `<rect x="${this.x}" y="${this.y}" width="${this.width}" height="${this.height}"/><text font-family="Calibri" x="60" y="80" font-size="60" font-weight="bold" fill="white" stroke= "black" stroke-width="3" text-anchor="middle">`
+        return ` width="300" height="200" viewBox="10 10 100 100" xmlns="http://www.w3.org/2000/svg"><rect x="${this.x}" y="${this.y}" width="${this.width}" height="${this.height}"/><text font-family="Calibri" x="60" y="80" font-size="55" font-weight="bold" fill="white" stroke= "black" stroke-width="3" text-anchor="middle">`
     }
 }
 
@@ -36,7 +36,7 @@ class Triangle {
         this.pgp6 = pgp6;
     }
     render(){
-        return `<polygon points="${this.pgp1}, ${this.pgp2} ${this.pgp3}, ${this.pgp4} ${this.pgp5}, ${this.pgp6}"/><text font-family="Calibri" x="150" y="145" font-size="75" font-weight="bold" fill="white" stroke= "black" stroke-width="3" text-anchor="middle">`
+        return ` width="300" height="200" viewBox="50 5 200 200" xmlns="http://www.w3.org/2000/svg"><polygon points="${this.pgp1}, ${this.pgp2} ${this.pgp3}, ${this.pgp4} ${this.pgp5}, ${this.pgp6}"/><text font-family="Calibri" x="150" y="145" font-size="75" font-weight="bold" fill="white" stroke= "black" stroke-width="3" text-anchor="middle">`
     }
 }
 
@@ -46,18 +46,20 @@ constructor(d){
     this.d = d;
 }
 render(){
-    return `<path d="${this.d}" /><text font-family="Calibri" x="50" y="60" font-size="40" font-weight="bold" style="stroke: black;" fill="#FFF" text-anchor="middle">`
+    return ` width="300" height="200" viewBox="-45 -50 200 200" xmlns="http://www.w3.org/2000/svg">
+    <path d="${this.d}" /><text font-family="Calibri" x="50" y="60" font-size="40" font-weight="bold" 
+    style="stroke: black;" fill="#FFF" text-anchor="middle">`
 }
 }
-
-// skull <g transform="translate(0.000000,200.000000) scale(0.100000,-0.100000)"><path d="M2928 1925 c-3 -3 -54 -11 -114 -19 -126 -16 -238 -60 -409 -163 -55 -32 -122 -68 -150 -78 -75 -28 -146 -104 -169 -181 -10 -34 -29 -74 -43 -88 -13 -15 -77 -62 -142 -105 -65 -44 -134 -98 -154 -121 -61 -72 -57 -135 14 -195 27 -22 38 -25 68 -20 60 11 171 75 240 139 l33 31 -8 -50 c-4 -27 -8 -65 -8 -83 0 -26 -1 -30 -9 -16 -5 9 -7 37 -3 62 l6 46 -27 -26 c-16 -14 -52 -41 -82 -59 -42 -25 -58 -43 -77 -83 -13 -28 -24 -57 -24 -64 0 -31 -43 -136 -62 -152 -37 -30 -214 -50 -363 -40 -87 6 -162 -50 -153 -113 3 -23 17 -29 113 -46 33 -6 80 -17 105 -24 l45 -14 -30 -13 c-16 -6 -68 -14 -115 -17 -99 -7 -151 -29 -180 -76 -21 -34 -26 -82 -11 -106 6 -10 19 -12 42 -7 18 3 84 11 146 16 66 6 135 18 165 30 61 23 61 23 53 10 -11 -18 -170 -60 -228 -60 -53 0 -58 -2 -68 -27 -13 -35 6 -87 29 -78 9 3 62 10 118 16 69 6 147 23 246 53 223 68 213 62 258 158 45 96 90 143 200 211 42 25 83 54 93 64 9 11 59 104 112 207 52 104 112 213 131 242 53 79 131 144 247 204 56 29 104 54 106 56 8 7 40 186 50 279 13 132 24 305 18 305 -2 0 -6 -2 -9 -5z m-908 -1025 c-50 -58 -81 -84 -43 -37 66 82 79 97 86 97 5 0 -14 -27 -43 -60z m-60 -73 c0 -2 -10 -12 -22 -23 l-23 -19 19 23 c18 21 26 27 26 19z m-50 -41 c0 -2 -8 -10 -17 -17 -16 -13 -17 -12 -4 4 13 16 21 21 21 13z"/><path d="M1465 1769 c-110 -11 -408 -100 -444 -132 -13 -12 -40 -58 -61 -103 -40 -85 -65 -111 -170 -175 -100 -60 -128 -87 -161 -157 -17 -37 -68 -136 -112 -221 -114 -217 -167 -271 -355 -366 l-89 -45 -21 -123 c-22 -124 -46 -412 -37 -428 18 -28 145 -19 253 18 68 22 259 119 315 159 26 19 72 43 100 54 99 37 149 89 181 188 24 72 50 100 169 177 221 145 264 234 160 325 -29 26 -42 31 -67 26 -51 -10 -147 -62 -213 -115 -35 -28 -63 -45 -63 -38 1 6 7 39 15 72 l15 60 -5 -54 -4 -54 22 20 c12 11 49 38 83 61 66 44 89 82 114 190 16 68 48 115 85 123 117 28 212 39 302 33 88 -5 101 -3 131 15 39 25 57 60 48 95 -5 20 -14 26 -44 31 -20 3 -74 14 -118 25 -104 25 -210 25 -320 0 -99 -23 -126 -19 -38 5 33 9 66 19 74 22 15 5 157 18 315 29 50 3 104 10 122 15 46 15 93 72 93 114 0 61 -5 65 -71 58 -32 -3 -59 -3 -59 0 0 3 5 18 12 32 8 19 8 31 0 50 -13 28 -11 28 -157 14z m3 -106 c-16 -2 -40 -2 -55 0 -16 2 -3 4 27 4 30 0 43 -2 28 -4z m-403 -515 c-50 -41 -60 -52 -115 -122 -44 -57 -76 -83 -88 -72 -3 3 0 6 5 6 6 0 39 33 73 73 56 66 144 147 158 147 4 0 -11 -15 -33 -32z m-536 -968 c-42 -22 -174 -60 -204 -59 -11 0 19 11 66 24 47 13 101 30 120 39 49 21 60 19 18 -4z"/></g>
 
 class Hands {
     constructor(d){
         this.d = d;
     }
     render(){
-        return `<g transform="${this.d}<text font-family="Calibri" x="150" y="120" font-size="55" font-weight="bold" fill="white" style="stroke: black;" text-anchor="middle">`
+        return ` width="300" height="200" viewBox="48 0 200 200" xmlns="http://www.w3.org/2000/svg">
+        <g transform="${this.d}<text font-family="Calibri" x="150" y="120" font-size="55" font-weight="bold" 
+        fill="white" style="stroke: black;" text-anchor="middle">`
     }
 }
 
@@ -66,7 +68,9 @@ class Skull {
         this.d = d;
     }
     render(){
-        return `<g transform="${this.d}<text font-family="Calibri" x="270" y="555" font-size="155" font-weight="bold" fill="white" stroke= "black" stroke-width="5" text-anchor="middle">`
+        return ` width="300" height="200" viewBox="50 0 500 700" xmlns="http://www.w3.org/2000/svg">
+        <g transform="${this.d}<text font-family="Calibri" x="270" y="555" font-size="155" font-weight="bold" 
+        fill="white" stroke= "black" stroke-width="5" text-anchor="middle">`
     }
 }
 
@@ -75,7 +79,9 @@ class Bird {
         this.d = d;
     }
     render(){
-        return `<g transform="${this.d}<text font-family="Calibri" x="170" y="135" font-size="55" font-weight="bold" fill="white" style="stroke: black;" text-anchor="middle">`
+        return ` width="300" height="200" viewBox="50 0 500 700" xmlns="http://www.w3.org/2000/svg">
+        <g transform="${this.d}<text font-family="Calibri" x="170" y="135" font-size="55" font-weight="bold" fill="white" 
+        style="stroke: black;" text-anchor="middle">`
     }
 }
 module.exports = { Circle, Square, Triangle, Heart, Hands, Skull, Bird };
